requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
httpx>=0.27.0
aiohttp>=3.10.0
pandas>=2.2.0
openpyxl>=3.1.2
faiss-cpu>=1.8.0
tiktoken>=0.7.0
pydantic>=2.9.0
python-dotenv>=1.0.0
playwright>=1.47.0
PyYAML>=6.0.2
beautifulsoup4>=4.12.3
numpy>=2.0.0
lxml>=5.3.0
dashscope>=1.20.0